requirements.txt
# Core
openai==2.31.0
numpy==2.2.6
python-dotenv==1.2.2

# Phase 1 — LangChain + ChromaDB
langchain==1.2.15
langchain-openai==1.1.12
langchain-community==0.4.1
chromadb==1.5.7

# Phase 2 — LangGraph
langgraph==1.1.6

# Phase 3 — API + UI
fastapi==0.135.3
uvicorn==0.44.0
streamlit==1.56.0
requests==2.33.1
